move mock data into pages: mockExecution & mockStepLogs
